--- src/main/scala/io/gustavoamigo/quill/pgsql/encoding/range/RangeSupport.scala
@@ -1,7 +1,0 @@
-package io.gustavoamigo.quill.pgsql.encoding.range
-
-import io.getquill.source.jdbc.JdbcSource
-
-trait RangeSupport  extends Decoders with Encoders {
-  this: JdbcSource[_, _] =>
-}
