+++ src/main/scala/io/gustavoamigo/quill/pgsql/encoding/range/numeric/NumericRangeSupport.scala
@@ -1,0 +1,7 @@
+package io.gustavoamigo.quill.pgsql.encoding.range.numeric
+
+import io.getquill.source.jdbc.JdbcSource
+
+trait NumericRangeSupport extends Decoders with Encoders {
+  this: JdbcSource[_, _] =>
+}
